Remove Kotlin JVM toolchain configuration
This commit removes the Kotlin JVM toolchain setup from the build.gradle.kts file. The `jvmToolchain` configuration, along with source and target compatibility settings, have been deleted to run according to the pipeline settings. It might make more sense to leave this to the build settings in the individual project. I'll revisit later.